Move callbacks for setting DQN epsilon values to the library
