+++ examples/atari/atari_iqn_hl.py
@@ -3,10 +3,6 @@
 import os
 from collections.abc import Sequence
 
-from examples.atari.atari_callbacks import (
-    TestEpochCallbackDQNSetEps,
-    TrainEpochCallbackNatureDQNEpsLinearDecay,
-)
 from examples.atari.atari_network import (
     IntermediateModuleFactoryAtariDQN,
 )
@@ -17,6 +13,10 @@ from tianshou.highlevel.experiment import (
     IQNExperimentBuilder,
 )
 from tianshou.highlevel.params.policy_params import IQNParams
+from tianshou.highlevel.trainer import (
+    TrainerEpochCallbackTestDQNSetEps,
+    TrainerEpochCallbackTrainDQNEpsLinearDecay,
+)
 from tianshou.utils import logging
 from tianshou.utils.logging import datetime_tag
 
@@ -84,9 +84,9 @@ def main(
         )
         .with_preprocess_network_factory(IntermediateModuleFactoryAtariDQN(features_only=True))
         .with_trainer_epoch_callback_train(
-            TrainEpochCallbackNatureDQNEpsLinearDecay(eps_train, eps_train_final),
+            TrainerEpochCallbackTrainDQNEpsLinearDecay(eps_train, eps_train_final),
         )
-        .with_trainer_epoch_callback_test(TestEpochCallbackDQNSetEps(eps_test))
+        .with_trainer_epoch_callback_test(TrainerEpochCallbackTestDQNSetEps(eps_test))
         .with_trainer_stop_callback(AtariStopCallback(task))
         .build()
     )
