+++ tianshou/utils/logger/base.py
@@ -7,7 +7,7 @@ from typing import Any
 
 import numpy as np
 
-VALID_LOG_VALS_TYPE = int | Number | np.number | np.ndarray
+VALID_LOG_VALS_TYPE = int | Number | np.number | np.ndarray | float
 VALID_LOG_VALS = typing.get_args(
     VALID_LOG_VALS_TYPE,
 )  # I know it's stupid, but we can't use Union type in isinstance
